fix: 코드래빗 리뷰 반영 수정

+++ src/main/java/com/team3/monew/repository/InterestRepositoryCustom.java
@@ -8,6 +8,6 @@ public interface InterestRepositoryCustom {
 
   List<Interest> searchByCondition(InterestSearchCondition condition);
 
-  int countByCondition(InterestSearchCondition condition);
+  long countByCondition(InterestSearchCondition condition);
 
 }
